move ap_platform.py to Archipelago/Platform.py
